+++ packages/labs/analyzer/src/test/analyzer_test.ts
@@ -4,51 +4,56 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
+import 'source-map-support/register.js';
 import {suite} from 'uvu';
 // eslint-disable-next-line import/extensions
 import * as assert from 'uvu/assert';
 import * as path from 'path';
 import {fileURLToPath} from 'url';
+import {getOutputFilename, getSourceFilename, languages} from './utils.js';
 
-import {Analyzer} from '../lib/analyzer.js';
-import {AbsolutePath} from '../lib/paths.js';
+import {createPackageAnalyzer, Analyzer, AbsolutePath} from '../index.js';
 
-const test = suite<{analyzer: Analyzer; packagePath: AbsolutePath}>(
-  'Basic Analyzer tests'
-);
-
-test.before((ctx) => {
-  try {
-    const packagePath = (ctx.packagePath = fileURLToPath(
-      new URL('../test-files/basic-elements', import.meta.url).href
-    ) as AbsolutePath);
-    ctx.analyzer = new Analyzer(packagePath);
-  } catch (error) {
-    // Uvu has a bug where it silently ignores failures in before and after,
-    // see https://github.com/lukeed/uvu/issues/191.
-    console.error('uvu before error', error);
-    process.exit(1);
-  }
-});
-
-test('Reads project files', ({analyzer, packagePath}) => {
-  const rootFileNames = analyzer.programContext.program.getRootFileNames();
-  assert.equal(rootFileNames.length, 5);
-
-  const elementAPath = path.resolve(packagePath, 'src', 'element-a.ts');
-  const sourceFile =
-    analyzer.programContext.program.getSourceFile(elementAPath);
-  assert.ok(sourceFile);
-});
-
-test('Analyzer finds class declarations', ({analyzer}) => {
-  const result = analyzer.analyzePackage();
-  const elementAModule = result.modules.find(
-    (m) => m.sourcePath === path.normalize('src/class-a.ts')
+for (const lang of languages) {
+  const test = suite<{analyzer: Analyzer; packagePath: AbsolutePath}>(
+    `Basic Analyzer tests (${lang})`
   );
-  assert.equal(elementAModule?.jsPath, path.normalize('out/class-a.js'));
-  assert.equal(elementAModule?.declarations.length, 1);
-  assert.equal(elementAModule?.declarations[0].name, 'ClassA');
-});
 
-test.run();
+  test.before((ctx) => {
+    try {
+      const packagePath = (ctx.packagePath = fileURLToPath(
+        new URL(`../test-files/${lang}/basic-elements`, import.meta.url).href
+      ) as AbsolutePath);
+      ctx.analyzer = createPackageAnalyzer(packagePath);
+    } catch (error) {
+      // Uvu has a bug where it silently ignores failures in before and after,
+      // see https://github.com/lukeed/uvu/issues/191.
+      console.error('uvu before error', error);
+      process.exit(1);
+    }
+  });
+
+  test('Reads project files', ({analyzer, packagePath}) => {
+    const rootFileNames = analyzer.program.getRootFileNames();
+    assert.equal(rootFileNames.length, 5);
+
+    const elementAPath = path.resolve(
+      packagePath,
+      getSourceFilename('element-a', lang)
+    );
+    const sourceFile = analyzer.program.getSourceFile(elementAPath);
+    assert.ok(sourceFile);
+  });
+
+  test('Analyzer finds class declarations', ({analyzer}) => {
+    const result = analyzer.getPackage();
+    const elementAModule = result.modules.find(
+      (m) => m.sourcePath === getSourceFilename('class-a', lang)
+    );
+    assert.equal(elementAModule?.jsPath, getOutputFilename('class-a', lang));
+    assert.equal(elementAModule?.declarations.length, 1);
+    assert.equal(elementAModule?.declarations[0].name, 'ClassA');
+  });
+
+  test.run();
+}