+++ packages/labs/analyzer/test-files/ts/modules/src/module-a.ts
@@ -1,0 +1,10 @@
+/**
+ * @license
+ * Copyright 2022 Google LLC
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+import {a, b, c} from './module-b.js';
+import {LitElement} from 'lit';
+
+export {a, b, c, LitElement};