+++ packages/labs/cli/CHANGELOG.md
@@ -1,9 +1,20 @@
 # @lit-labs/cli
 
-## 0.0.1
+## 0.2.0
+
+### Minor Changes
+
+- [#3304](https://github.com/lit/lit/pull/3304) [`31bed8d6`](https://github.com/lit/lit/commit/31bed8d6542c44a64bad8282b9ce5e5d6514e44a) - Added support for analyzing JavaScript files.
+
+- [#2936](https://github.com/lit/lit/pull/2936) [`7a9fc0f5`](https://github.com/lit/lit/commit/7a9fc0f57e43c2eab44e9442e5896f951a8c751a) - Locally version and lazily install the localize command.
 
 ### Patch Changes
 
-- Updated dependencies []:
-  - @lit-labs/gen-utils@0.0.1
-  - @lit-labs/gen-wrapper-react@0.0.1
+- Updated dependencies [[`31bed8d6`](https://github.com/lit/lit/commit/31bed8d6542c44a64bad8282b9ce5e5d6514e44a), [`569a6237`](https://github.com/lit/lit/commit/569a6237377eeef0c8dced2c369c77ebdd81218e), [`fc2fd4c8`](https://github.com/lit/lit/commit/fc2fd4c8f4a25b9a85073afcb38614209e079bb9)]:
+  - @lit-labs/analyzer@0.3.0
+
+## 0.1.0
+
+### Minor Changes
+
+- [#3225](https://github.com/lit/lit/pull/3225) [`198da7ce`](https://github.com/lit/lit/commit/198da7ceabc944b142a666cae56ea239624cd019) - Initial release